Refactor replay buffer (mode out from deepq to commons)
